Add requirement for pythonX-jinja2
This change is adding the requirement pythonX-jinja2 to the tripleo-ansible
package. This is being done because we require jinja2 to use some of our
filters and we can not rely on assumed dependencies to bring in the right,
most up to date, versions.

Change-Id: I51140423832abf5e2728077eb2993c69ee38f243
Signed-off-by: Kevin Carter <kecarter@redhat.com>
(cherry picked from commit 21f300ae7024af312c655274aaef43a6a9f57124)