Update patch set 1
Patch Set 1:

I'm not sure this is exactly the right place to do this, so hoping someone can correct me if needs be.

The tripleo_collectd[1] and tripleo_metrics_qdr[2] roles in tripleo-ansible depend on our underlying config roles which are packaged seperately. Right now, I need to do `dnf install ansible-role-collectd-config ansible-role-qdr-config` on my undercloud in order to get these packages in place. We'd like them to be included by default so we don't have to document that step.

Is this the right way to do it?

[1] https://review.opendev.org/plugins/gitiles/openstack/tripleo-ansible/+/refs/heads/master/tripleo_ansible/roles/tripleo_collectd/tasks/configure_collectd.yml#219

[2] https://review.opendev.org/plugins/gitiles/openstack/tripleo-ansible/+/refs/heads/master/tripleo_ansible/roles/tripleo_metrics_qdr/tasks/create_config.yml#130

Patch-set: 1